Rename CSSSCheckerEngineTests to CSSSCheckerEngine.Tests
