+++ CSSSCheckerEngine.Tests/CSSSCheckerEngine.Tests.csproj
@@ -1,0 +1,55 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<Project DefaultTargets="Build" ToolsVersion="4.0" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
+  <PropertyGroup>
+    <Configuration Condition=" '$(Configuration)' == '' ">Debug</Configuration>
+    <Platform Condition=" '$(Platform)' == '' ">AnyCPU</Platform>
+    <ProjectGuid>{40A0EAD2-7A6B-4C02-9315-5AEF401AFD24}</ProjectGuid>
+    <OutputType>Library</OutputType>
+    <RootNamespace>CSSSCheckerEngineTests</RootNamespace>
+    <AssemblyName>CSSSCheckerEngineTests</AssemblyName>
+    <TargetFrameworkVersion>v4.0</TargetFrameworkVersion>
+    <ReleaseVersion>0.3</ReleaseVersion>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Debug|AnyCPU' ">
+    <DebugSymbols>true</DebugSymbols>
+    <DebugType>full</DebugType>
+    <Optimize>false</Optimize>
+    <OutputPath>bin\Debug</OutputPath>
+    <DefineConstants>DEBUG;</DefineConstants>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+  </PropertyGroup>
+  <PropertyGroup Condition=" '$(Configuration)|$(Platform)' == 'Release|AnyCPU' ">
+    <Optimize>true</Optimize>
+    <OutputPath>bin\Release</OutputPath>
+    <ErrorReport>prompt</ErrorReport>
+    <WarningLevel>4</WarningLevel>
+  </PropertyGroup>
+  <ItemGroup>
+    <Reference Include="System" />
+    <Reference Include="nunit.framework, Version=3.12.0.0, Culture=neutral, PublicKeyToken=2638cd05610744eb, processorArchitecture=MSIL">
+      <HintPath>..\packages\NUnit.3.12.0\lib\net40\nunit.framework.dll</HintPath>
+    </Reference>
+    <Reference Include="Microsoft.CSharp" />
+  </ItemGroup>
+  <ItemGroup>
+    <Compile Include="IssueFilesTests.cs" />
+    <Compile Include="Issues\Files\Contents.Files.Tests.cs" />
+    <Compile Include="Issues\Files\Existence.Files.Tests.cs" />
+    <Compile Include="Issues\System\Registry.System.Tests.cs" />
+  </ItemGroup>
+  <ItemGroup>
+    <None Include="packages.config" />
+  </ItemGroup>
+  <ItemGroup>
+    <ProjectReference Include="..\CSSSCheckerEngine\CSSSCheckerEngine.csproj">
+      <Project>{7DF38F4C-6B96-4E2F-A76C-8F21D976CE56}</Project>
+      <Name>CSSSCheckerEngine</Name>
+    </ProjectReference>
+    <ProjectReference Include="..\CSSSConfig\CSSSConfig.csproj">
+      <Project>{2B005BD2-677F-4646-BB17-673AD50DF07D}</Project>
+      <Name>CSSSConfig</Name>
+    </ProjectReference>
+  </ItemGroup>
+  <Import Project="$(MSBuildBinPath)\Microsoft.CSharp.targets" />
+</Project>
