+++ CSSSCheckerEngine.Tests/Issues/System/Registry.System.Tests.cs
